requirements.txt
datasets>=2.18.0
PyYAML>=6.0.1
tqdm>=4.66.0
torch>=2.1.0
transformers>=4.41.0,<5.0.0
sentence-transformers>=5.3.0
accelerate>=1.0.0
faiss-cpu>=1.13.0
tensorflow>=2.21.0,<2.22.0
tf-keras>=2.21.0,<2.22.0
beir>=2.2.0
mlflow>=2.20.0

# Optional for lexical-only workflows.
rank-bm25>=0.2.2
pytrec-eval-terrier>=0.5.6
